requirements.txt
argparse
opencv-python
Pillow
pandas
numpy
imgaug
tensorflow-cpu
tensorflow-addons